Move conftest.py up
This feels a bit wrong but I'd rather be able to run pytest against the main
qtrio package for now.  Perhaps we'll find a better solution later.

https://docs.pytest.org/en/latest/writing_plugins.html#testing-plugins

________________________ ERROR collecting test session _________________________
Defining 'pytest_plugins' in a non-top-level conftest is no longer supported:
It affects the entire test suite instead of just below the conftest as expected.
  /opt/hostedtoolcache/Python/3.8.3/x64/lib/python3.8/site-packages/qtrio/_tests/conftest.py
Please move it to a top level conftest file at the rootdir:
  /opt/hostedtoolcache/Python/3.8.3/x64/lib/python3.8/site-packages/qtrio
For more information, visit:
  https://docs.pytest.org/en/latest/deprecations.html#pytest-plugins-in-non-top-level-conftest-files